+++ src/module.ts
@@ -1,32 +1,57 @@
-import {defineNuxtModule, addPlugin, createResolver, addTemplate} from '@nuxt/kit'
-
-// Module options TypeScript interface definition
-export interface ModuleOptions {
-  firstName : string;
-}
+import {defineNuxtModule, addPlugin, createResolver, addTemplate, addImports, addImportsDir} from '@nuxt/kit'
+import ModuleOptions from './interface/moduleOptions';
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
-    name: 'my-nuxt3-sample',
-    configKey: 'ghanavati',
+    name: 'capModule',
+    configKey: 'CapModule',
   },
   // Default configuration options of the Nuxt module
   defaults: {
-    firstName:""
+    db_name:'',
+    client_id:'',
+    is_multi_token:false,
+    environment:'Development',
+    api_methods : {
+      user_info: '',
+      authorization_by_app_code: ''
+    },
+    production : {
+      base_url:'',
+      redirect_uri:'',
+      sso_site_url:'',
+    },
+    development : {
+      base_url:'',
+      redirect_uri:'',
+      sso_site_url:'',
+    }
   },
   setup(options, _nuxt) {
+    //@ts-ignore
     const resolver = createResolver(import.meta.url)
 
-    _nuxt.options.alias['#well-known'] = addTemplate({
-      filename: 'well-known.mjs',
+    _nuxt.options.alias['#capModule'] = addTemplate({
+      filename: 'cap-oauth.mjs',
       write: true,
       getContents: () => `export default ${JSON.stringify(options, undefined, 2)}`
     }).dst || ''
 
-    // Do not add the extension since the `.ts` will be transpiled to `.mjs` after `npm run prepack`
+    //#region Register composables
+    // Register composables directory
+    addImportsDir(resolver.resolve('./runtime/composables'))
+    //#endregion
+
+    //#region Register plugins
     addPlugin({
-      src: resolver.resolve('./runtime/plugin'),
+      src: resolver.resolve('./runtime/config.client'),
       mode: 'all',
     })
+    addPlugin({
+        src: resolver.resolve('./runtime/plugin'),
+        mode: 'all',
+      })
+    //#endregion
+
   },
 })