+++ src/runtime/plugin.ts
@@ -1,10 +1,10 @@
-import {defineNuxtPlugin} from '#app'
-
+//#region Import
 // @ts-ignore
-import wellKnownOptions from '#well-known'
+import {defineNuxtPlugin} from '#app'
+// @ts-ignore
+import CapModule from '#capModule'
+//#endregion
 
 export default defineNuxtPlugin(({ nuxtApp, $config }) => {
-
-  console.log('Plugin injected by my-module!')
-  console.log('STEP 2 : ',wellKnownOptions)
+  console.log('Plugin injected by CAP!')
 })