+++ CHANGELOG.md
@@ -1,5 +1,19 @@
-# Changelog
+# CAP Modules
+Auth , API Caller , IndexedDB
+
+Developed By : [CAP Company](https://i-cap.ir) , Ahmad Ghanavati
 
 
-## v0.0.4
-
+## Change Log :
+## v0.0.1
+- add Login  function
+- add logout function
+- add userInfo function
+- add authorizationByAppCode function
+- add api function
+- add api with credentials function
+- add indexed db
+- add indexed db insert Data function
+- add indexed db get Data function
+- add indexed db remove Data function
+- add indexed db clear Data function